Extract updateFavorite to own utils file
